moved to floats

+++ pte-tree-mxj/src/main/java/triangularsneaky/tree/vision/pte/attentionHoggers/WriteableMatrix.java
@@ -1,5 +1,5 @@
 package triangularsneaky.tree.vision.pte.attentionHoggers;
 
 public interface WriteableMatrix extends Matrix {
-    void set(int x, int y, int[] values);
+    void set(int x, int y, double[] values);
 }
